requirements.txt
<<<<<<< HEAD
Flask==2.3.3
requests==2.31.0
python-dotenv==1.0.0
=======
PyGithub==2.3.0
python-dotenv==1.0.0
markdown2==2.4.10
Flask==3.0.2
pdfkit==1.0.0
python-docx==1.1.0
matplotlib==3.8.2
seaborn==0.13.0
plotly==5.18.0
requests==2.31.0
>>>>>>> 54089dd76f0f3700af83b8e07eecec5aebed2144
